Add SILAttr but don't wire it up.
A SILAttr is a struct consisting of:

1. A SILAttrKind.
2. A StringRef or Integer Value.

A SILAttrKind describes the properties of the enum via various metaprogramming
routines. For instance, one can ask if a SILAttr with SILAttrKind Foo can have
other SILAttr with the same SILAttrKind in the same attribute list. Some
examples of SILAttrKinds are:

1. Function_Semantics: A function level semantic string attribute.
2. Function_GlobalInit: A function level global init boolean attribute.
3. Return_Semantics: A string semantic attribute on a return value.
4. Parameter_Semantics: A string semantic attribute on a parameter.

*NOTE* The last two examples are hypothetical. I have only added function level
attributes fow now.

Every SILAttrKind is mapped to a SILAttrValueKind which is an enum consisting of
one of: String, Integer, IntegerSet. These value kind's define the behavior of a
SILAttr of SILAttrKind in a Function's Attribute List. Specifically:

1. A SILAttr with a String type can occur multiple times in an attribute list
with different strings. Each string value can only occur once in a list and the
strings will be uniqued in the SILModule. This is for things like @semantics.
2. A SILAttr with an Integer type can occur only once in an attribute list. This
represents both boolean values and integral enums which can only legally take on
one value. This is for things like GlobalInit, InlineStrategy.
3. A SILAttr with an IntegerSet type can occur multiple times in an attribute
list. This represents enums which can have multiple values in an attribute list
at the same time, i.e. Effects.