[semantic-arc] As staging detail, add a HasQualifiedOwnership flag on all SILFunctions.
I thought a bit over the past day or two about how we are going to need to manage the
EnableSemanticARC state as we move the SemanticARC flag backwards
through the SILOptimizer pipeline. There are a few things that came of
these thoughts:

1. The Eliminator pass really needs to be a function pass to ensure that
we can transparently put it at any stage of the optimization
pipeline. This means we need a bit on all SILFunctions to state whether
or not a SILFunction has ownership with all functions initially having
ownership enabled.

2. We do not need to store that bit when printing or serializing SIL by
following the following logic: There are three sets of instructions in
SIL from an ownership perspective:

a. ownership qualified instructions
b. ownership unqualified instructions
c. instructions that do not have ownership semantics (think literals, geps, etc).

The set of functions can be split into ownership qualified and ownership
unqualified using the rules that:

a. a function can never contain both ownership qualified and ownership
unqualified instructions.
b. a function that contains only instructions without ownership
semantics is considered ownership qualified.

Thus we can know when parsing/serializing what category of function we
have and set the bit appropriately.

rdar://28685236