[semantic-sil] When emitting a direct argument for a callsite, if the value is not guaranteed after we call getAsSingleValue(), borrow the value.
The reason why we do not want to do this in getAsSingleValue itself is that
getAsSingleValue is asking the question of, can the argument source provide the
object at +0. We are handling the case of whether or not the argument source
providing the object at +0, we need it at +0, so create a begin_borrow context.