[sil] Add new APIs SILBuilder::createAllocRef{,Dynamic}() that take a SmallVectorImpl.
In the constructor for AllocRef{,Dynamic}Inst::create(...), we already
copied ArrayRef<...> into llvm::SmallVector<...>. This moves this change into
SILBuilder for the ArrayRef case and exposes the underlying llvm::SmallVector
case. This ensures that we do not create two SmallVectorImpls at the cost of an
additional iteration to collect conformances. An extra iteration is always
cheaper than mallocing memory.