[semantic-sil] Add an optional PostOrderFunctionInfo argument to SILValue::verifyOwnership().
When we run this code from the verifier, we will always reuse the same
PostOrderFunctionInfo, so in most cases this will not be a nullptr and no slow
down will result. But the nullptr will allow for this code to be run from the
debugger without the user having to create their own PostOrderFunctionInfo.

This commit does not contain the code that wires this up to the actual ownership
verifier.