[ownership-verifier] Teach the ownership verifier how to handle no return functions from SILGen in Raw SIL.
SILGen does not place in unreachables after emitting code for no return
functions. This is to help diagnostics. This does cause issues since for non-try
apply no-return functions, we need to emit the cleanups /before/ the no-return.

In the try_apply case, due to the way the instruction is setup, we still have a
place to put the destroys even though technically in the no-return case we will
be leaking memory if the no-return doesn't ever call the catch path.