[ownership-verifier] When begin_borrow is applied to an address, it produces a trivial value.
Even though we do not support borrow verification of addresses in the SIL
Ownership Model, sometimes due to the recursive nature of SILGen, address only
types will have a begin_borrow applied. For now, just tell the verifier that
this is producing a trivial value.

Once AddressOnlyTypes are no longer represented as addresses, I expect this to
go away and we should be able to tighten this up.