[README.md] Add explicit references to the CI requiring committer access and how to become a committer.
I have seen over time people confused about what the requirement is for being
able to trigger swift-ci testing. This commit attempts to alleviate this issue
by adding a section to the top level README of swift that:

1. Calls out CI directly via a new subsection called 'Continuous Integration on
Github'.
2. Explicitly says in the section that one must be a committer to trigger the CI
or trigger merging. Then a link to the CI docs are provided as well as a direct
link to the gaining commit access subsection of the Contributors guide.