[silgen] Split SILGenBuilder into two builders: ConformanceEndowedSILBuilder and SILGenBuilder.
For a while SILGenBuilder has been two things jammed into one:

1. A ManagedValue trafficing class that attempted to handle ownership transfer
for users.
2. A SILBuilder subclass that override specific SILInstruction creating
instructions so that we properly "endowed" the instructions with those
conformances.

By having these in one class (instead of using composition), we made it possible
for one to by mistake invoke a SILBuilder instead of a SILGenBuilder method for
one of those instructions causing us to not properly endow the instruction with
its conformances.