[ownership] Update the ownership utils infrastructure to treat mark_dependence as forwarding.
We have already been forwarding ownership in terms of ValueOwnership and
OwnershipUtils, I just had not setup certain parts of the ownership utils to
recognize mark_dependence as forwarding of guaranteed values. We did not hit
this before since we have not had been late enough in the pipeline to get
mark_dependence on guaranteed values. I am fixing a problem with closure
lifetime fixup that exposed this issue.