[inlining] Fix inlining/devirt of coroutines in ossa.
There were two bugs here:

* Devirt: I fixed the inserting of end_borrows when replacing a begin_apply with
a devirt begin_apply target. The problem is that we were setting the wrong
insertion point and inserting the end_borrows /before/ the user, rather than
after the token user.

* Normal inlining: If we had an owned value and used it as aguaranteed argument,
when inlining we need to insert a begin_borrow/end_borrow around the newly
inlined region. Previously we created the values before inlining resulting in
memory errors. Instead, now we put in the begin_borrow before we inline and then
after all inlining and cloning have been completed and our region is finalized,
insert the end_borrows. This ensures that we do not need to worry about how the
CFG is changed during inlining to get the right scope.