[swift-siloptimizer-pass] Add a Package.swift/xcconfig file for quick prototyping with Swift SILOptimizer passes.
This enables one to play around with this in Xcode/etc. The reason why we need
the xcconfig is to ensure that Xcode is able to find the SwiftBridge module in
./swift/include/swift-c/SwiftBridge. Currently this works well enough for
building, but SourceKit will not generate a header for SwiftBridge for some
reason.

To generate an xcodeproj:

swift package generate-xcodeproj --xcconfig-overrides Package.xcconfig