[closure-lifetime-fixup] Use GetValueAtEndOfBlock instead of GetValueInMiddleOfBlock for getting values in exits.
Otherwise we will leak values if the copy block is in the same block as the
terminator and we fail to identify a single destroy for the copy block. We are
generally pretty good at identifying those single destroy cases, so I am not
sure how often this actually happens.

Found by the ownership verifier.