[build-script-impl] Make --install-llvm an alias for setting LLVM_INSTALL_COMPONENTS to all.
LLVM is the only project with the behavior that it is triggered by
LLVM_INSTALL_COMPONENTS being sent to a non-empty value rather than
INSTALL_LLVM. This preserves what is currently supported but also allows for
build-script-impl to be standard here.

I need this so that I can add support in build-script for --install-all.