Revert "build-script: remove dead CMake options for Swift"
This reverts commit 95102bc258a6b30ba2349d561ef56ea48c601cae.

This is actually a dead option that is relied upon for some configurations. I am
going to add a note to not touch the option.

I talked with compnerd about this and they are cool with this.

I think that we can redo this in a nicer way when we are further into the
build-script-impl refactoring.