[ownership] project_box is an interior pointer vending inst so make it only accept guaranteed values as operands.
This enforces that the project_box's users must be within the guaranteed value's
borrow scope or the ownership verifier will assert. Most importantly this is the
last interior pointer vending instruction that is not guarded by a borrow scope
meaning that now we can shrink lifetimes in SemanticARCOpts without worrying
about shrinking lifetimes past uses of an interior pointer, consider:

```
%0 = alloc_box
%1 = project_box %0
... uses of %1 ...
destroy_value %0
```

In this example, there is no connection in between the uses of %1 (since it is
an address) and %0 itself, so just looking at SSA edges, there is no reason that
an optimizer would know without analyzing the project_box's users how far it
could shrink the lifetime of the object. In contrast, by the IR requiring a
borrow scope and statically guaranteeing that all users of `%1` are within the
scope, we can do a simple SSA def-use check to figure out how much we can shrink
the lifetime of the destroy_value.