[sil] Add a dylib that the pass manager in NDEBUG builds can weakly link against and then support verification of SILModules, SILFunctions during optimization.
The idea is that:

1. We already always ship the frontend option sil-verify-all, it just doesn't do
anything today when NDEBUG is set since we compile out the underlying
verification routines.

2. Given that is true, why not use a weak symbol and the driver's ability to
re-run frontend codegen jobs that crash to instead of just compiling out the
verification to change NDEBUG builds to instead check for the weak pointer and
if it is set, call it for verification. This would then cause -sil-verify-all to
work if the driver when it reran the job to inject the weak symbol.

In the case of the library not being enabled, the cost is just checking if a
function pointer is null which should be cheap. There would be no other
runtime/other cost during normal running meaning that we can use this mechanism
to enable more aggressive verification.

One thing to be aware is that the dylib would be rev-locked to the compiler and
is technically a part of the compiler.