[ownership] Add a rebase_borrow instruction to enable elimination of ARC in loops using induction on type shapes.
This instruction can be used to represent that a borrowed value with base
oldBase is not depending on a different base newBase for its liveness
requirement. With that in mind, we treat the rebase_borrow as an instantaneous
use of its first parameter and a borrow use of its second parameter, the new
base.

In order to preserve safety we require that:

1. If oldBase is an object, newBase must be its original base.
2. If oldBase is an address, then oldBase must be an interior pointer whose base
is newBase. We require that any interior pointers from newBase must never write
to memory in the rebase_borrow region.

With both of those conditions, we can guarantee that the oldBase if it is an
object or an address remains undisturbed while newBase is around.