[sil] Refactor out the serialization of TwoOperandLayout insts into a helper struct.
Just cleaning up the unmaintainable function that is writeSILInstruction and
trying to show a manner in which we can start to refactor out the work in each
switch case into separate helper code. With that in hand, we should be able to
eventually change this giant switch into a visitor which is easier to reason
about as separate pieces of code.